Endret barnetilsynsutgift til bigdecimal

+++ src/main/kotlin/no/nav/bidrag/bidragskalkulator/dto/åpenBeregning/ÅpenBeregningRequestDto.kt
@@ -9,6 +9,7 @@ import jakarta.validation.constraints.NotEmpty
 import jakarta.validation.constraints.NotNull
 import no.nav.bidrag.bidragskalkulator.dto.*
 import no.nav.bidrag.domene.enums.beregning.Samværsklasse
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Schema(description = "Informasjon om et barn i beregningen")
@@ -28,7 +29,7 @@ data class BarnMedAlderDto(
     override val bidragstype: BidragsType,
 
     @Schema(description = "Utgifter i kroner per måned som den bidragsmottaker har til barnetilsyn for dette barnet", required = false)
-    override val barnetilsynsutgift: Double?
+    override val barnetilsynsutgift: BigDecimal?
 ): IFellesBarnDto {
     @JsonIgnore
     @Schema(hidden = true) // Hides from Swagger
